Move the new Tests (from release5.9) into another file - move new tests to their file - rename select.test to db_index.test to use it for index tests
