+++ test/models/userRefs.ts
@@ -1,14 +1,14 @@
 import { arrayProp, prop, Ref, Typegoose } from '../../src/typegoose';
 
 export class UserRef extends Typegoose {
-    @prop({ ref: UserRef, default: null })
-    public master?: Ref<UserRef>;
+  @prop({ ref: UserRef, default: null })
+  public master?: Ref<UserRef>;
 
-    @arrayProp({ itemsRef: UserRef, default: [] })
-    public subAccounts!: Ref<UserRef>[];
+  @arrayProp({ itemsRef: UserRef, default: [] })
+  public subAccounts!: Ref<UserRef>[];
 
-    @prop({ required: true })
-    public name!: string;
+  @prop({ required: true })
+  public name!: string;
 }
 
 export const UserRefModel = new UserRef().getModelForClass(UserRef);
